[FIX] fixed clone_oca_dependencies when repo url is not specified
When cloning the dependencies inside a `oca_dependencies.txt` file that
doesn't have URLs, but only the dependency name...

Before this: the dependencies were cloned by default using the
organization owner of the repository of the build, even when the owner
of the `oca_dependencies.txt` file was different.

Now: if the `oca_dependencies.txt` only specifies the repository name,
the dependencies will be cloned using the same owner as the owner of the
dependencies file.

For example: If you have a repository owned by `vauxoo` organization and
that repo has a dependency of another repo owned by `oca` (like
`stock-logistics-warehouse`) that in turn has an `oca_dependencies.txt`
file with no URLs, those dependencies are going to be gotten from `oca`
as well.

---
Closes https://github.com/Vauxoo/maintainer-quality-tools/issues/215