[FIX] fixed getaddons.py to print the repos by dependency level
When loading the addons from a folder to create the `addons_path`...

Before this: the addons are loaded as they are found in the folder, with
no specific order, and no regards for the dependencies.

Now: the addons are loaded following the level of the dependency.

For example: considering `vauxoo/imeqmo` as a base repository, that repo
has a dependency of `vauxoo/stock-logistics-warehouse` which in turn has
a dependency of `vauxoo/stock-logistics-tracking` and
`vauxoo/stock-logistics-barcode` (in that order), the final
`addons_path` will respect their dependency level and will result in
`imeqmo,warehouse,tracking,barcode` (repos paths abbreviated).